+++ src/main/java/ru/mal/unialbumsbackend/domain/requests/CreateAlbumRequest.java
@@ -11,13 +11,13 @@ public class CreateAlbumRequest {
 
     private String cover;
 
-    private double tracksRating;
+    private int tracksRating;
 
-    private double atmosphereRating;
+    private int atmosphereRating;
 
-    private double bitsRating;
+    private int bitsRating;
 
-    private double textRating;
+    private int textRating;
 
     private String artist;
